refactor: change project directories feat: add webpack module feat: add prov and dev scripts with webpack
