+++ package.json
@@ -5,8 +5,8 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-    "start:dev": "nodemon server.js"
+    "start:prod": "webpack --mode production && node ./dist/bundle",
+    "start:dev": "nodemon ./src/server.js"
   },
   "repository": {
     "type": "git",
@@ -22,5 +22,9 @@
     "dotenv": "^10.0.0",
     "nodemon": "^2.0.15",
     "uuid": "^8.3.2"
+  },
+  "devDependencies": {
+    "webpack": "^5.64.4",
+    "webpack-cli": "^4.9.1"
   }
 }
